app_local/requirements.txt
absl-py==1.0.0
answerquest @ git+https://github.com/roemmele/answerquest.git@d2a11ba414d6673d15ac14c66b34bb6c1f3944b4
blessed==1.19.1
blis==0.4.1
cachetools==5.0.0
catalogue==1.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
ConfigArgParse==1.5.3
cycler==0.11.0
cymem==2.0.6
dominate==2.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
filelock==3.6.0
Flask==2.1.1
Flask-Bootstrap4==4.0.2
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-WTF==1.0.1
fonttools==4.33.2
future==0.18.2
gevent==21.12.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.44.0
gunicorn==20.1.0
huggingface-hub==0.5.1
idna==3.3
importlib-metadata==4.11.3
inquirer==2.9.2
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
murmurhash==1.0.6
nlp2==1.8.43
nlp2go==0.4.12
nltk==3.5
numpy==1.19.1
oauthlib==3.2.0
OpenNMT-py==1.2.0
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
plac==1.1.3
preshed==3.0.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyngrok==5.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-editor==1.0.4
pytorch-crf==0.7.2
pytz==2022.1
pywsd==1.2.4
PyYAML==6.0
rank-bm25==0.2.1
readchar==3.0.5
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.96
six==1.16.0
sklearn==0.0
spacy==2.3.2
srsly==1.0.5
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tfkit==0.7.21
thinc==7.4.1
threadpoolctl==3.1.0
tokenizers==0.8.1rc1
tomli==2.0.1
torch==1.6.0
torchtext==0.4.0
tqdm==4.64.0
transformers==3.0.2
typing_extensions==4.2.0
urllib3==1.26.9
visitor==0.1.3
waitress==2.1.1
wasabi==0.9.1
wcwidth==0.2.5
Werkzeug==2.1.1
wn==0.0.22
WTForms==3.0.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
